requirements.txt
Flask==2.2.3
flask_cors
python-dotenv
faiss-cpu
sentence-transformers
groq
PyPDF2
pydub
yt_dlp
beautifulsoup4
google-genai
google-generativeai
gunicorn==20.0.4
grpcio
Werkzeug==2.2.3